# src/controllers/chatController.ts
import { Request, Response, NextFunction } from "express";
import asyncHandler from "express-async-handler";

import { Chat } from "../models/chatModel";
import { User } from "../models/userModel";

//@description     Create or fetch One to One Chat
//@route           POST /api/chats/create-chat
//@access          Protected
const createChat = asyncHandler(
  async (req: Request, res: Response, next: NextFunction) => {
    const { userId } = req.body as { userId: string };

    if (!userId) {
      return res.status(400).json({ error: "UserId not sent with request" });
    }

    let isChat = await Chat.find({
      isGroupChat: false,
      $and: [
        { users: { $elemMatch: { $eq: req.user._id } } },
        { users: { $elemMatch: { $eq: userId } } },
      ],
    })
      .populate("users", "-password")
      .populate("latestMessage");

    isChat = await User.populate(isChat, {
      path: "latestMessage.sender",
      select: "name picture email",
    });

    if (isChat.length > 0) {
      res.send(isChat[0]);
    } else {
      let chatData = {
        chatName: "sender",
        isGroupChat: false,
        users: [req.user._id, userId],
      };

      try {
        const createdChat = await Chat.create(chatData);
        const FullChat = await Chat.findOne({ _id: createdChat._id }).populate(
          "users",
          "-password"
        );
        res.status(200).json(FullChat);
      } catch (error: any) {
        res.status(400);
        throw new Error(error.message);
      }
    }
  }
);

//@description     Fetch all chats for a user
//@route           GET /api/chats/fetch-chats
//@access          Protected
const fetchChats = asyncHandler(
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      Chat.find({ users: { $elemMatch: { $eq: req.user._id } } })
        .populate("users", "-password")
        .populate("groupAdmin", "-password")
        .populate("latestMessage")
        .sort({ updatedAt: -1 })
        .then(async (results) => {
          results = await User.populate(results, {
            path: "latestMessage.sender",
            select: "mobile_number picture email",
          });
          res.status(200).send(results);
        });
    } catch (error: any) {
      res.status(400);
      throw new Error(error.message);
    }
  }
);

//@description     Create New Group Chat
//@route           POST /api/chats/create-group-chat
//@access          Protected
const createGroupChat = asyncHandler(
  async (req: Request, res: Response, next: NextFunction) => {
    if (!req.body.users || !req.body.name) {
      return res.status(400).send({ message: "Please Fill all the feilds" });
    }

    var users = JSON.parse(req.body.users);

    if (users.length < 2) {
      return res
        .status(400)
        .send("More than 2 users are required to form a group chat");
    }

    users.push(req.user);

    try {
      const groupChat = await Chat.create({
        chatName: req.body.name,
        users: users,
        isGroupChat: true,
        groupAdmin: req.user,
      });

      const fullGroupChat = await Chat.findOne({ _id: groupChat._id })
        .populate("users", "-password")
        .populate("groupAdmin", "-password");

      res.status(200).json(fullGroupChat);
    } catch (error: any) {
      res.status(400);
      throw new Error(error.message!);
    }
  }
);

// @desc    Rename Group
// @route   PUT /api/chats/update-group-name
// @access  Protected
const updateGroupName = asyncHandler(
  async (req: Request, res: Response, next: NextFunction) => {
    const { chatId, chatName } = req.body;

    const updatedChat = await Chat.findByIdAndUpdate(
      chatId,
      {
        chatName: chatName,
      },
      {
        new: true,
      }
    )
      .populate("users", "-password")
      .populate("groupAdmin", "-password");

    if (!updatedChat) {
      res.status(404);
      throw new Error("Chat Not Found");
    } else {
      res.json(updatedChat);
    }
  }
);

// @desc    Remove user from Group
// @route   PUT /api/chats/remove-from-group
// @access  Protected
const removeFromGroup = asyncHandler(
  async (req: Request, res: Response, next: NextFunction) => {
    const { chatId, userId } = req.body;

    // check if the requester is admin

    const removed = await Chat.findByIdAndUpdate(
      chatId,
      {
        $pull: { users: userId },
      },
      {
        new: true,
      }
    )
      .populate("users", "-password")
      .populate("groupAdmin", "-password");

    if (!removed) {
      res.status(404);
      throw new Error("Chat Not Found");
    } else {
      res.json(removed);
    }
  }
);

// @desc    Add user to Group / Leave
// @route   PUT /api/chats/add-to-group
// @access  Protected
const addToGroup = asyncHandler(
  async (req: Request, res: Response, next: NextFunction) => {
    const { chatId, userId } = req.body;

    // check if the requester is admin

    const added = await Chat.findByIdAndUpdate(
      chatId,
      {
        $push: { users: userId },
      },
      {
        new: true,
      }
    )
      .populate("users", "-password")
      .populate("groupAdmin", "-password");

    if (!added) {
      res.status(404);
      throw new Error("Chat Not Found");
    } else {
      res.json(added);
    }
  }
);

exports = {
  createChat,
  fetchChats,
  createGroupChat,
  updateGroupName,
  addToGroup,
  removeFromGroup,
};